corpvideohost/requirements.txt
asgiref==3.7.2
attrs==23.2.0
boto3==1.34.69
botocore==1.34.69
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==25.2.0
geckodriver-autoinstaller==0.1.0
googleapis-common-protos==1.63.0
graphviz==0.20.3
grpcio==1.62.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
jmespath==1.0.1
outcome==1.3.0.post0
packaging==24.0
pillow==10.2.0
pluggy==1.5.0
protobuf==4.25.3
psycopg2==2.9.9
pycparser==2.21
pydotplus==2.0.2
PyJWT==2.8.0
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.2.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
selenium==4.21.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2023.4
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
yandexcloud==0.267.0
